# swagger.yml
openapi: 3.0.1

info:
  title: Portfolio Budget
  description: Simple backend API to manage portfolio budget using an envelope budgeting
    method
  version: 1.0.0


servers:
- url: https://portfolio-budget-starter.herokuapp.com/
- url: http://localhost:4001/


paths:
  /api/v1/envelopes:
    get:
      tags:
      - Envelopes
      summary: Get all envelopes
      responses:
        200:
          description: Returns a list of all envelopes
          content: {}
    post:
      tags:
      - Envelopes
      summary: Make a new envelop
      requestBody:
        description: Write a new envelope
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Envelop'
      responses:
        201:
          description: A new envelop successufully made
          content: {}
        500:
          description: Internal server error
          content: {}
  /api/v1/envelopes/{id}:
    get:
      tags:
      - Envelopes
      summary: Get a specific envelop by ID
      parameters:
      - name: id
        in: path
        description: Enter a number to fetch an envelop
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Returns an envelope along with its data
          content: {}
        404:
          description: User not found
          content: {}
        500:
          description: Internal server error
          content: {}
    put:
      tags:
      - Envelopes
      summary: Update a specific envelop
      produces:
        - application/json
      parameters:
      - name: id
        in: path
        description: Fetch the envelop you want to update
        required: true
      requestBody:
        description: Write a new envelope
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Envelop'
      responses:
        201:
          description: Envelop updated
          content: {}
        500:
          description: Internal server error
          content: {}
    delete:
      tags:
      - Envelopes
      summary: Delete an envelop
      produces:
        - application/json
      parameters:
      - name: id
        in: path
        description: Delete an envelop
        required: true
        example: 1
      responses:
        204:
          description: Successfully deleted the envelop
          content: {}
        500:
          description: Internal server error
          content: {}
  /api/v1/envelopes/{fromId}/transfer/{toId}:
    post:
      tags:
        - Envelopes
      summary: Transfer money
      produces:
        - application/json
      parameters:
      - name: fromId
        in: path
        description: fromId
        required: true
      - name: toId
        in: path
        description: toId
        required: true
      requestBody:
        description: Transfer the money from fromId to toId
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
              example:
                amount: 200
      responses:
        201:
          description: Amount of cash has been sent
          content: {}
        500:
          description: Not enough money or someother error
          content: {}

components:
  schemas:
    Envelop:
      type: object
      required:
        - title
        - budget
      properties:
        title:
          type: string
        budget:
          type: integer
      example:
        title: Pay up punk
        budget: 3000